requirements.txt
pyautogui
Pillow
mss
img2pdf